Fixed coverflow crash when covers are moving and device is rotated.
coverViews are initially set to NSNull and lazily assigned to
TKCoverflowCoverView as required. The code in adjustViewHeirarchy was
not checking for NSNull when calling sendSubviewToBack causing a crash.
This bug was discovered by scrolling through the covers and rotating
the device. Fix is to check for NSNull before calling sendSubviewToback.